完善功能
如果想抓取搜狐http://w.sohu.com网页，并且获得css，js和图片文件，可以通过python main.py -d 60 -u
http://w.sohu.com -o /tmp/backup

+++ main.py
@@ -1,4 +1,4 @@
-#-*-coding:utf-8-*-
+#coding=utf-8
 import urllib2
 import re
 import os
@@ -6,10 +6,12 @@ from time import strftime, localtime
 import time
 from bs4 import BeautifulSoup 
 import shutil  
+import sys,getopt
 
-
-def getpic(url,content,lists):
-    os.mkdir("/tmp/backup"+lists+"/images")
+#获取图像文件
+def getpic(url,lists):
+    content  = urllib2.urlopen(url).read()
+    os.makedirs(con+"/"+lists+"/images")
     soup = BeautifulSoup(content)  
     imgs = soup.findAll('img')
 
@@ -22,17 +24,23 @@ def getpic(url,content,lists):
         imgUrls = re.findall(r'.*src=\"([^\"]*)\"',str(img))
         if imgNames and imgType and imgUrls:
        
-            imagecontent = urllib2.urlopen(imgUrls[0]).read() 
-            imageName = imgNames[0]+"."+imgType[0]    
-            
-            with open(imageName,'wb') as code:
-                code.write(imagecontent)
-                code.close()
-            shutil.move(imageName,"/tmp/backup"+lists+"/images")
+            imagecontent = urllib2.urlopen(imgUrls[0]).read() #读进image数据
+            imageName = imgNames[0]+"."+imgType[0]    #image文件的名字
+            print imageName
+            filename =con+ "/"+lists+"/images/"+imageName
+         #   print filename
+            if os.path.exists(filename): 
+                print ""
+            else:
+                with open(imageName,'wb') as code:
+                    code.write(imagecontent)
+                    code.close()
+                shutil.move(imageName,con+"/"+lists+"/images")
 
-
-def getcss(url,content,lists):
-    os.mkdir("/tmp/backup"+lists+"/css")
+#获取css文件
+def getcss(url,lists):
+    content  = urllib2.urlopen(url).read()
+    os.makedirs(con+"/"+lists+"/css")
     soup = BeautifulSoup(content,from_encoding="utf-8")  
     csss = soup.findAll('link',attrs={'type':'text/css'})  
 
@@ -41,56 +49,72 @@ def getcss(url,content,lists):
         cssurls = re.findall(r'.*href=\"([^\"]*)\"',str(css))  
 
         if cssnames: 
-            csscontent = urllib2.urlopen(cssurls[0]).read() 
-            cssName = cssnames[0]+".css"     
+            csscontent = urllib2.urlopen(cssurls[0]).read() #读进css数据
+            cssName = cssnames[0]+".css"      #css文件的名字
             
             with open(cssName,'wb') as code:
                 code.write(csscontent)
                 code.close()
-            shutil.move(cssName,"/tep/backup"+lists+"/css")
+            shutil.move(cssName,con+"/"+lists+"/css")
                 
 
-
-def getjs(url,content,lists):
-    os.mkdir("/tmp/backup"+lists+"/js")
+#获取js文件
+def getjs(url,lists):
+    content  = urllib2.urlopen(url).read()
+    os.makedirs("/tmp/backup/"+lists+"/js")
     soup = BeautifulSoup(content,from_encoding="utf-8")  
-    jss = soup.findAll('script',attrs={'type':'text/javascript'})  
+    jss = soup.findAll('script',attrs={'type':'text/javascript'})  #找出含有type="text/javascript"特性的 
     for js in jss:
         jsnames = re.findall(r'.*/(.*)\.js',str(js)) 
         jsurls = re.findall(r'.*src=\"([^\"]*)\"',str(js))  
 
         if jsnames and jsurls:  
-            jscontent = urllib2.urlopen(jsurls[0]).read() 
-            jsName = jsnames[0]+".js"      
+            jscontent = urllib2.urlopen(jsurls[0]).read() #读进js数据
+            jsName = jsnames[0]+".js"      #js文件的名字
             with open(jsName,'w+') as code:
                 code.write(jscontent)
                 code.close()
-            shutil.move(jsName,"/tmp/backup"+lists+"/js")
-
-def gethtml(url,content,lists):
-  #  html_name = index.html
-    
+            shutil.move(jsName,con+"/"+lists+"/js")
+
+#获取index.html文件
+def gethtml(url,lists):
+    content  = urllib2.urlopen(url).read()
     fp = open('index.html',"w+")
     fp.write(content)
     fp.close()
-    shutil.move('index.html','/tep/backup'+lists)
+    shutil.move('index.html',con+'/'+lists)
 
 
-             
-year = strftime("%Y",localtime()) 
-mon  = strftime("%m",localtime()) 
-day  = strftime("%d",localtime()) 
-hour = strftime("%H",localtime()) 
-mins  = strftime("%M",localtime())
-
+#获取系统时间信息
+option,long = getopt.getopt(sys.argv[1:],"hd:u:o:")
+for key,value in option:
+    if key in ("-d"):
+        get_time = int(value)
+    if key in ("-u"):
+        get_url = value
+    if key in ("-o"):
+        con = value
+print get_time
+print get_url
+print con
 while True:
-    print year+mon+day+hour+mins
-    lists = year+mon+day+hour+mins 
-    os.mkdir('/tmp/backup/'+lists)
-    url = 'http://m.sohu.com'
+    year = strftime("%Y",localtime()) 
+    mon  = strftime("%m",localtime()) 
+    day  = strftime("%d",localtime()) 
+    hour = strftime("%H",localtime()) 
+    mins  = strftime("%M",localtime())
+
+
+#   print year+mon+day+hour+mins
+    lists = year+mon+day+hour+mins
+    os.makedirs(con+"/"+lists)
+
+    
+    url = get_url
+    
     content  = urllib2.urlopen(url).read()
-    gethtml(url,content,lists)
-    getjs(url,content,lists)
-    getcss(url,content,lists)
-    getpic(url,content,lists)
-    time.sleep(60)
+    gethtml(url,lists)
+    getjs(url,lists)
+    getcss(url,lists)
+    getpic(url,lists)
+    time.sleep(get_time)
